@@ -1,8 +1,8 @@
-import React from "react"
-import ReactDOM from "react-dom"
+import { createRoot } from "react-dom/client"
 
 import App from "./children/App"
 
-const root = document.getElementById("root")
+const container = document.getElementById("root")
 
-ReactDOM.render(<App />, root)
+const root = createRoot(container)
+root.render(<App />)
