@@ -3,17 +3,19 @@ import JsonFetcher from "./JsonFetcher"
 
 class BooksDisplay extends React.Component {
   render() {
-    const book_rows = this.props.data && this.props.data.map(book => {
-      const subject_list = book.subjects && book.subjects.join("; ")
-      return (
-        <tr key={book.key}>
-          <td>{book.title}</td>
-          <td>{book.subtitle}</td>
-          <td>{subject_list}</td>
-          <td>{book.publish_date}</td>
-        </tr>
-      )
-    })
+    const book_rows =
+      this.props.data &&
+      this.props.data.map((book) => {
+        const subject_list = book.subjects && book.subjects.join("; ")
+        return (
+          <tr key={book.key}>
+            <td>{book.title}</td>
+            <td>{book.subtitle}</td>
+            <td>{subject_list}</td>
+            <td>{book.publish_date}</td>
+          </tr>
+        )
+      })
 
     return (
       <div className="panel panel-default">
@@ -22,9 +24,7 @@ class BooksDisplay extends React.Component {
         </div>
         <div className="panel-body">
           <table className="table table-striped">
-            <tbody>
-              {book_rows}
-            </tbody>
+            <tbody>{book_rows}</tbody>
           </table>
         </div>
       </div>
@@ -34,7 +34,8 @@ class BooksDisplay extends React.Component {
 
 class Books extends React.Component {
   render() {
-    const url = "http://openlibrary.org/query.json?type=/type/edition&limit=10&*=&title=" +
+    const url =
+      "http://openlibrary.org/query.json?type=/type/edition&limit=10&*=&title=" +
       encodeURIComponent(this.props.term)
 
     return <JsonFetcher DataConsumer={BooksDisplay} url={url} />
