@@ -3,20 +3,23 @@ import JsonFetcher from "./JsonFetcher"
 
 class RedditDisplay extends React.Component {
   render() {
-    const reddit_links = this.props.data && this.props.data.data && this.props.data.data.children.map(
-      ({ data: { id, title, url, created, subscribers } }) => {
-        const date = new Date(1000 * created) // Convert seconds to milliseconds
-        return (
-          <tr key={id}>
-            <td>{date.toLocaleDateString()}</td>
-            <td>{subscribers}</td>
-            <td>
-              <a href={"https://reddit.com" + url}>{title}</a>
-            </td>
-          </tr>
-        )
-      }
-    )
+    const reddit_links =
+      this.props.data &&
+      this.props.data.data &&
+      this.props.data.data.children.map(
+        ({ data: { id, title, url, created, subscribers } }) => {
+          const date = new Date(1000 * created) // Convert seconds to milliseconds
+          return (
+            <tr key={id}>
+              <td>{date.toLocaleDateString()}</td>
+              <td>{subscribers}</td>
+              <td>
+                <a href={"https://reddit.com" + url}>{title}</a>
+              </td>
+            </tr>
+          )
+        }
+      )
 
     return (
       <div className="panel panel-default">
@@ -35,7 +38,8 @@ class RedditDisplay extends React.Component {
 
 class Reddit extends React.Component {
   render() {
-    const url = "https://api.reddit.com/api/subreddit_autocomplete_v2.json?limit=10&include_over_18=false&query=" +
+    const url =
+      "https://api.reddit.com/api/subreddit_autocomplete_v2.json?limit=10&include_over_18=false&query=" +
       encodeURIComponent(this.props.term)
 
     return <JsonFetcher DataConsumer={RedditDisplay} url={url} />
