@@ -1,17 +1,19 @@
 import React from "react"
 
-const BooksDisplay = props => {
-  const book_rows = props.data && props.data.map(book => {
-    const subject_list = book.subjects && book.subjects.join("; ")
-    return (
-      <tr key={book.key}>
-        <td>{book.title}</td>
-        <td>{book.subtitle}</td>
-        <td>{subject_list}</td>
-        <td>{book.publish_date}</td>
-      </tr>
-    )
-  })
+const BooksDisplay = (props) => {
+  const book_rows =
+    props.data &&
+    props.data.map((book) => {
+      const subject_list = book.subjects && book.subjects.join("; ")
+      return (
+        <tr key={book.key}>
+          <td>{book.title}</td>
+          <td>{book.subtitle}</td>
+          <td>{subject_list}</td>
+          <td>{book.publish_date}</td>
+        </tr>
+      )
+    })
 
   return (
     <div className="panel panel-default">
@@ -20,9 +22,7 @@ const BooksDisplay = props => {
       </div>
       <div className="panel-body">
         <table className="table table-striped">
-          <tbody>
-            {book_rows}
-          </tbody>
+          <tbody>{book_rows}</tbody>
         </table>
       </div>
     </div>
@@ -49,12 +49,12 @@ class Books extends React.Component {
   fetchBookData() {
     fetch(
       "http://openlibrary.org/query.json?type=/type/edition&limit=10&*=&title=" +
-      encodeURIComponent(this.props.term)
+        encodeURIComponent(this.props.term)
     )
-      .then(response => {
+      .then((response) => {
         return response.json()
       })
-      .then(json => {
+      .then((json) => {
         this.setState({ data: json })
       })
   }
