@@ -7,11 +7,11 @@ const dev = process.env.NODE_ENV !== "production"
 const HTMLWebpackPluginConfig = new HTMLWebpackPlugin({
   template: path.join(__dirname, "/src/index.html"),
   filename: "index.html",
-  inject: "body"
+  inject: "body",
 })
 
 const DefinePluginConfig = new webpack.DefinePlugin({
-  "process.env.NODE_ENV": JSON.stringify("production")
+  "process.env.NODE_ENV": JSON.stringify("production"),
 })
 
 module.exports = {
@@ -19,9 +19,9 @@ module.exports = {
     host: "0.0.0.0",
     port: "3500",
     headers: {
-      "Access-Control-Allow-Origin": "*"
+      "Access-Control-Allow-Origin": "*",
     },
-    historyApiFallback: true
+    historyApiFallback: true,
   },
   devtool: "source-map",
   entry: [path.join(__dirname, "/src/index.jsx")],
@@ -30,19 +30,19 @@ module.exports = {
       {
         test: /\.jsx?$/,
         exclude: /node_modules/,
-        loaders: ["babel-loader"]
-      }
-    ]
+        use: "babel-loader",
+      },
+    ],
   },
   resolve: {
-    extensions: [".js", ".jsx"]
+    extensions: [".js", ".jsx"],
   },
   output: {
     filename: "index.js",
-    path: path.join(__dirname, "/build")
+    path: path.join(__dirname, "/build"),
   },
   mode: dev ? "development" : "production",
   plugins: dev
     ? [HTMLWebpackPluginConfig]
-    : [HTMLWebpackPluginConfig, DefinePluginConfig]
+    : [HTMLWebpackPluginConfig, DefinePluginConfig],
 }
